Create README.md
Esg Dao 
our focus.  -  finding and investing in projects that deliver tax advantaged financial returns along with social and environmental benefits to the communities that they exist in. The larger goal being decarbonization at scale as a byproduct of becoming a De-centralized power company to the coastal communities we serve a De-centralized power company to the coastal communities we serve

The mission of The Scubaverse is to bring more awarenesss and stakeholder engagement to Life Below Water (UN Sustainable development Goal #14), by leveraging web3 technology to gamify the experience of conservation and sustainable use of marine resources.